Revert "Moved UserPageFrame and UserChangeBudgetFrame to correct package"
