+++ src/main/java/Frames/UserPageFrame.java
@@ -1,4 +1,6 @@
-package ui;
+package Frames;
+
+import ui.OnboardingFrame;
 
 import javax.swing.*;
 import java.awt.event.ActionEvent;
@@ -13,7 +15,6 @@ public class UserPageFrame extends JFrame{
     private static JLabel userBudgetLabel;
     private static JLabel userCurrentBudgetLabel;
     private static JButton changeBudgetButton;
-    private static JButton viewPastOrdersButton;
     private static JButton backButton;
 
     public UserPageFrame() {
@@ -67,24 +68,14 @@ public class UserPageFrame extends JFrame{
             @Override
             public void actionPerformed(ActionEvent e) {
                 exit();
-                ui.UserChangeBudgetFrame userChangeBudgetFrame = new UserChangeBudgetFrame();
+                UserChangeBudgetFrame userChangeBudgetFrame = new UserChangeBudgetFrame();
             }
         });
         panel.add(changeBudgetButton);
 
-        viewPastOrdersButton = new JButton("View Past Orders");
-        viewPastOrdersButton.setBounds(10, 175, 180, 25);
-        viewPastOrdersButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                exit();
-                PastOrdersFrame pastOrdersFrame = new PastOrdersFrame();
-            }
-        });
-        panel.add(viewPastOrdersButton);
 
         backButton = new JButton("Back");
-        backButton.setBounds(10, 205, 80, 25);
+        backButton.setBounds(10, 175, 80, 25);
         backButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
