--- src/main/java/ui/PastOrdersFrame.java
@@ -1,4 +1,0 @@
-package ui;
-
-public class PastOrdersFrame {
-}
